Spliting some functions
